examples/run-20230923-043300_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/whatsapp/0/LLMatic/services/message_service.py
import sqlite3

DB_NAME = 'chat.db'

class MessageService:
	@staticmethod
	def send_message(sender, receiver, message):
		conn = sqlite3.connect(DB_NAME)
		c = conn.cursor()
		c.execute('INSERT INTO messages (sender, receiver, message, read_receipt, encrypted) VALUES (?, ?, ?, ?, ?)', (sender, receiver, message, 0, 0))
		message_id = c.lastrowid
		conn.commit()
		conn.close()
		return {'id': message_id, 'sender': sender, 'receiver': receiver, 'message': message, 'read_receipt': 0, 'encrypted': 0}

	@staticmethod
	def queue_message(receiver, message):
		pass

	@staticmethod
	def send_queued_messages():
		pass
